+++ src/main/kotlin/snyk/iac/IacIssue.kt
@@ -1,6 +1,5 @@
 package snyk.iac
 
-import com.google.gson.annotations.Expose
 import io.snyk.plugin.Severity
 
 data class IacIssue(
@@ -14,48 +13,10 @@ data class IacIssue(
     val impact: String,
     val resolve: String? = null,
     val references: List<String> = emptyList(),
-    val path: List<String> = emptyList(),
-    @Expose val obsolete: Boolean = false
+    val path: List<String> = emptyList()
 ) {
     var ignored = false
+    var obsolete = false
 
     fun getSeverity(): Severity = Severity.getFromName(severity)
-
-    // We need the equals / hashcode methods as we don't want to include `obsolete` in our equality checks
-    @Suppress("DuplicatedCode") // false positive for generated code
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as IacIssue
-
-        if (id != other.id) return false
-        if (title != other.title) return false
-        if (severity != other.severity) return false
-        if (publicId != other.publicId) return false
-        if (documentation != other.documentation) return false
-        if (lineNumber != other.lineNumber) return false
-        if (issue != other.issue) return false
-        if (impact != other.impact) return false
-        if (resolve != other.resolve) return false
-        if (references != other.references) return false
-        if (path != other.path) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = id.hashCode()
-        result = 31 * result + title.hashCode()
-        result = 31 * result + severity.hashCode()
-        result = 31 * result + publicId.hashCode()
-        result = 31 * result + documentation.hashCode()
-        result = 31 * result + lineNumber
-        result = 31 * result + issue.hashCode()
-        result = 31 * result + impact.hashCode()
-        result = 31 * result + (resolve?.hashCode() ?: 0)
-        result = 31 * result + references.hashCode()
-        result = 31 * result + path.hashCode()
-        return result
-    }
 }