+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Fixed
 - All failed scan results for artifacts(file, image) are now shown in the results(Tree) too (right after successful scan results).
 - Open Source Scan results annotations are shown for Gradle(Groovy) dependencies.
+- Results (json) received from CLI are now sanitized for correctness.
 
 ### Added
 - Navigation to the Editor for the Open Source Scan results and all failed artifact's scan results.